requirement.txt
amqp==2.5.2
asgiref==3.2.3
autopep8==1.4.4
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
rope==0.14.0
some-package==0.1
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==1.0.0
